requirements.txt
fastapi>=0.100.0
uvicorn
motor
pydantic>=2.0.0
python-dotenv
email-validator
